Fixed more stuff in the Kitchen.

+++ Project_2/src/main/java/com/project2/kitchentable/controllers/KitchenController.java
@@ -1,8 +1,12 @@
 package com.project2.kitchentable.controllers;
 
 import java.util.UUID;
+import java.util.concurrent.ExecutionException;
 
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -10,7 +14,6 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.server.ServerWebExchange;
 
 import com.datastax.oss.driver.api.core.uuid.Uuids;
@@ -20,13 +23,18 @@ import com.project2.kitchentable.services.RecipeService;
 import com.project2.kitchentable.services.ReviewService;
 import com.project2.kitchentable.beans.Recipe;
 import com.project2.kitchentable.beans.Reviews;
+import com.project2.kitchentable.beans.User;
 
 import reactor.core.publisher.Mono;
 
 @RestController
 @RequestMapping(value = "/kitchen")
 public class KitchenController {
+	private static Logger log = LogManager.getLogger(AuthController.class);
 
+	
+	@Autowired
+	private AuthController authorize;
 	private KitchenService kitchenService;
 	private RecipeService recipeService;
 	private ReviewService reviewService;
@@ -47,52 +55,69 @@ public class KitchenController {
 	}
 
 	@PostMapping("/new")
-	public Mono<ResponseEntity<Kitchen>> addKitchen(@RequestBody Kitchen k) {
-		System.out.println("Making a new kitchen");
-		k.setId(Uuids.timeBased());
-		k.setHeadUser(Uuids.timeBased());
-		k.setFamilyID(Uuids.timeBased());
+	public Mono<ResponseEntity<Kitchen>> addKitchen(ServerWebExchange exchange, @RequestBody Kitchen k) {
+		User user = authorize.userAuth(exchange);
+		if(user == null) {
+			exchange.getResponse().setStatusCode(HttpStatus.BAD_REQUEST);
+			return null;
+		}		
+		k.setId(user.getKitchenID());
+		k.setHeadUser(user.getUserID());
+		k.setFamilyID(user.getFamilyID());
 		return kitchenService.addKitchen(k).map(kitchen -> ResponseEntity.status(201).body(kitchen))
 				.onErrorResume(error -> Mono.just(ResponseEntity.badRequest().body(k)));
 	}
 
 	@PostMapping("/addToList")
-	public Mono<ResponseEntity<Kitchen>> addToList(@RequestParam(name = "list", required = false) String listname,
-			@RequestParam(name = "kitchen", required = false) UUID kID,
+	public Mono<ResponseEntity<Kitchen>> addToList(ServerWebExchange exchange, @RequestParam(name = "list", required = false) String listname,
 			@RequestParam(name = "ingredient", required = false) UUID iID,
-			@RequestParam(name = "amount", required = false) Double amt) {
-		System.out.println("Adding to " + listname);
-		Kitchen k = kitchenService.getKitchenByID(kID).block();
-		return kitchenService.addFood(listname, k, iID, amt).map(kitchen -> ResponseEntity.status(201).body(kitchen))
-				.onErrorResume(error -> Mono.just(ResponseEntity.badRequest().body(k)));
+			@RequestParam(name = "amount", required = false) Double amt) throws InterruptedException, ExecutionException {
+		User user = authorize.userAuth(exchange);
+		if(user == null) {
+			exchange.getResponse().setStatusCode(HttpStatus.BAD_REQUEST);
+			return null;
+		}
+		log.debug(user.getKitchenID());
+		Kitchen k = kitchenService.getKitchenByID(user.getKitchenID()).toFuture().get();
+
+		return kitchenService.addFood(listname, k, iID, amt).map(kitchen -> ResponseEntity.status(201).body(kitchen)).onErrorStop();
+
+		
 	}
 
 	@GetMapping(value = "/removeFood")
-	public Mono<ResponseEntity<Kitchen>> removeFood(@RequestParam(name = "list", required = false) String listname,
-			@RequestParam(name = "kitchen", required = false) UUID kID,
+	public Mono<ResponseEntity<Kitchen>> removeFood(ServerWebExchange exchange, @RequestParam(name = "list", required = false) String listname,
 			@RequestParam(name = "ingredient", required = false) UUID iID,
 			@RequestParam(name = "amount", required = false) Double amt) throws Exception {
-		Kitchen k = kitchenService.getKitchenByID(kID).block();
+		User user = authorize.userAuth(exchange);
+		if(user == null) {
+			exchange.getResponse().setStatusCode(HttpStatus.BAD_REQUEST);
+			return null;
+		}
+		Kitchen k = kitchenService.getKitchenByID(user.getKitchenID()).block();
 		return kitchenService.removeFood(listname, k, iID, amt).map(kitchen -> ResponseEntity.status(201).body(kitchen))
 				.onErrorResume(error -> Mono.just(ResponseEntity.badRequest().body(k)));
 
 	}
 
 	@GetMapping(value = "/cook")
-	public Mono<ResponseEntity<String>> cook(@RequestParam(name = "recipe", required = true) UUID recipe,
-			@RequestParam(name = "kitchen", required = true) UUID kID, 
+	public Mono<ResponseEntity<String>> cook(ServerWebExchange exchange, @RequestParam(name = "recipe", required = true) UUID recipe,
 			@RequestParam(name = "review", required = false) String reviewBody,
-			@RequestParam(name = "score", required = false) Double score,
-			@RequestParam(name = "images", required = false) MultipartFile images) {
+			@RequestParam(name = "score", required = false) Double score) {
 		
-		return Mono.zip(kitchenService.getKitchenByID(kID), recipeService.getRecipeById(recipe)).flatMap(data -> {
+		User user = authorize.userAuth(exchange);
+		if(user == null) {
+			return Mono.just(ResponseEntity.badRequest().body("User is not logged in"));
+		}
+		
+		return Mono.zip(kitchenService.getKitchenByID(user.getKitchenID()), recipeService.getRecipeById(recipe)).flatMap(data -> {
 			Kitchen k = data.getT1();
 			Recipe r = data.getT2();
 
 			if((reviewBody != null) && (score != null)) {
 
 				return kitchenService.cook(r, k).flatMap(kitchen -> { // Attempt to cook 
-					Reviews rev = new Reviews(Uuids.timeBased(), Uuids.timeBased() /* This param will be changed to the logged in user's UUID */, recipe, score, reviewBody);
+					Reviews rev = new Reviews(Uuids.timeBased(), user.getUserID(), recipe, score, reviewBody);
 					return reviewService.addReview(rev).map(review -> ResponseEntity.status(201).body(kitchen.toString() + "\n" + review.toString())) // Proceed to try to add review
 						.onErrorResume(error -> Mono.just(ResponseEntity.badRequest().body(error.toString()))); // If error, assign error message to response
 				}).onErrorResume(error -> Mono.just(ResponseEntity.badRequest().body(error.toString())));
@@ -109,20 +134,28 @@ public class KitchenController {
 
 	@GetMapping(value = "/buyFood")
 	public Mono<ResponseEntity<Kitchen>> buyFood(ServerWebExchange exchange,
-			@RequestParam(name = "list", required = false) String listname,
-			@RequestParam(name = "kitchen", required = false) UUID kID,
 			@RequestParam(name = "ingredient", required = false) UUID iID,
 			@RequestParam(name = "amount", required = false) Double amt) {
-
-		Kitchen k = kitchenService.getKitchenByID(kID).block();
 		
-		if (k != null && k.getShoppingList().containsKey(iID)) {
-			Kitchen tempK = kitchenService.removeFood("shopping", k, iID, amt).block();
-			return kitchenService.addFood("inventory", tempK, iID, amt)
-					.map(kitchen -> ResponseEntity.status(201).body(kitchen))
-					.onErrorResume(error -> Mono.just(ResponseEntity.badRequest().body(k)));
+		User user = authorize.userAuth(exchange);
+		if(user == null) {
+			exchange.getResponse().setStatusCode(HttpStatus.BAD_REQUEST);
+			return null;
 		}
-		return Mono.just(ResponseEntity.badRequest().body(k));
+
+		return kitchenService.getKitchenByID(user.getKitchenID()).flatMap(k -> {
+			if (k != null && k.getShoppingList().containsKey(iID)) {
+				return kitchenService.removeFood("shopping", k, iID, amt).flatMap(tempK -> {
+					return kitchenService.addFood("inventory", tempK, iID, amt)
+							.map(kitchen -> ResponseEntity.status(201).body(kitchen))
+							.onErrorResume(error -> Mono.just(ResponseEntity.badRequest().body(k)));
+				});
+				
+			}
+			return Mono.just(ResponseEntity.badRequest().body(k));
+		});
+		
+
 	}
 
 }
